refactor(ProjectsForm): add common utils
Jira issue: VEGA-842
